add OrderDetail page

+++ DTOs/OrderDTO.cs
@@ -1,3 +1,5 @@
+using Microsoft.AspNetCore.Mvc;
+
 namespace Backend_CuoiKy.DTOs.Orders
 {
     public class OrderDTO
@@ -6,14 +8,14 @@ namespace Backend_CuoiKy.DTOs.Orders
         public int CustomerId { get; set; }
         public DateTime OrderDate { get; set; }
         public string? Status { get; set; }
-        public int TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
         public List<OrderDetailDTO>? Items { get; set; }
     }
     public class OrderDetailDTO
     {
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public int UnitPrice { get; set; }
-        public int SubTotal => Quantity * UnitPrice;
+        public decimal UnitPrice { get; set; }
+        public decimal SubTotal => Quantity * UnitPrice;
     }
 }
